Fail faster with OOME
TestThreadFailure test (and possibly others) are confused by a behavior
that repeatedly retries allocation requests rather than throwing OOME.
This change allows us to throw OOME if we have performed a full GC and
still cannot satisfy the request.

The problem with the existing code is that we can end up in an infinite
retry loop, if GC continues to "make progress", but the progress made is
not sufficient to satisfy the allocation request.